Add testAll
